Bug 1689620 - Rename auto_perf_sheriffing' s testing package
